LLM生成メッセージをPOST後表示
Next.jsサーバーサイド経由で、コンポーネントにメッセージを渡し、ブラウザでメッセージをバナー表示
