+++ front/src/app/calendar/page.tsx
@@ -11,6 +11,7 @@ import { EventContentArg } from "@fullcalendar/common";
 import Modal from "../_components/Modal";
 import { getFetchData, postFetchData, deleteFetchData } from "../fetch";
 import { EventInfo } from "../types";
+import MessageBanner from "../_components/MessageBannar"; // MessageBannerコンポーネントのパス
 
 let eventGuid = 0;
 
@@ -74,9 +75,7 @@ export default function Calendar() {
     startDate: string,
     endDate: string
   ) => {
-    const eventId = createEventId();
     const newEvent = {
-      id: eventId,
       title: eventTitle,
       start: startDate,
       end: endDate,
@@ -89,9 +88,13 @@ export default function Calendar() {
       const result = await postFetchData(eventTitle, startDate, endDate, 1);
       if (!result.error) {
         // POST処理が成功したら、カレンダーにイベントを追加する
-        const newEvents = [...events, newEvent];
+        // サーバーから返されたイベントIDを使用
+        const newEvents = [...events, { ...newEvent, id: result.id }];
         console.log("POST一覧", newEvents);
         setEvents(newEvents);
+        // バックエンドにイベントIDを送信してメッセージを生成する
+        setSelectedEventId(result.id); // MessageBannerに表示するIDをセット
+
         setEventTitle("");
         setSelectedDate("");
         setIsModalOpen(false);
@@ -170,9 +173,9 @@ export default function Calendar() {
   // };
 
   // 予定のIDを生成
-  const createEventId = () => {
-    return `event-${eventGuid++}`; // 予定のIDを生成
-  };
+  // const createEventId = () => {
+  //   return `event-${eventGuid++}`; // 予定のIDを生成
+  // };
 
   // モーダル閉じたときの処理
   const closeModal = () => {
@@ -221,6 +224,7 @@ export default function Calendar() {
           end: endDateTime,
         }}
       />
+      {selectedEventId && <MessageBanner id={selectedEventId} />}
     </>
   );
 }
